Restructure ZESPlus module
