+++ src/pages/ZESPlus-Research/sections/03_safety.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import { Heading, SectionProps } from '~/components2/Article';
+
+const SectionSafety = (props: SectionProps) => (
+  <>
+    <Heading as="h2" {...props}>
+      Auswertung der Sicherheit
+    </Heading>
+  </>
+);
+
+export default SectionSafety;
